requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
flatbuffers==23.5.26
fonttools==4.45.1
gast==0.5.4
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-vision==3.4.5
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h5py==3.10.0
httplib2==0.22.0
idna==3.6
imageio==2.33.0
imgaug==0.4.0
keras==2.15.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2023.9.26
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
Werkzeug==3.0.1
wrapt==1.14.1
